requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
scipy
python-dotenv
google-generativeai
streamlit